fix: remove unnecessary u32 cast in parser.rs
Remove redundant type cast from u32 to u32 in SourceLocation::new call.
This fixes the Clippy warning about unnecessary_cast.

+++ crates/sruja-language/src/parser.rs
@@ -134,7 +134,7 @@ impl Parser {
                     sruja_diagnostics::codes::CODE_SYNTAX_ERROR,
                     Severity::Error,
                     error_msg,
-                    SourceLocation::new(self.filename.clone(), error_pos as u32, 0),
+                    SourceLocation::new(self.filename.clone(), error_pos, 0),
                 )])
             }
         }
